vscode: downgrade nix-vscode-extensions
It turns out a typical cadence for breaking changes to the VS Code
plugin API is every two or three months. What the fuck.

Obviously this means it's not practical to pair nixos stable releases
with the rolling-release model of nix-vscode-extensions. So, I'm
downgrading nix-vscode-extensions to the commit that was live as of
24.05's freeze from nixpkgs master, and will just update it once every
six months, unless something else breaks.